# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Squeak Platform",
        "title": "Squeak API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.squeak.today",
    "basePath": "/",
    "paths": {
...
        "/audio/audiobook": {
            "get": {
                "description": "Get audiobook for a news_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audio"
                ],
                "summary": "Get audiobook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News ID",
                        "name": "news_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "story_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "story",
                        "name": "type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "1",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AudiobookResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/audio/stt": {
            "post": {
                "description": "Convert speech audio to text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audio"
                ],
                "summary": "Speech to text",
                "parameters": [
                    {
                        "description": "Speech to text request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SpeechToTextRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SpeechToTextResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/audio/translate": {
            "post": {
                "description": "Translate text from source language to target language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audio"
                ],
                "summary": "Translate text",
                "parameters": [
                    {
                        "description": "Translation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TranslateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TranslateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/audio/tts": {
            "post": {
                "description": "Convert text to speech audio",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "audio"
                ],
                "summary": "Text to speech",
                "parameters": [
                    {
                        "description": "Text to speech request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TextToSpeechRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TextToSpeechResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AudiobookResponse": {
            "type": "object",
            "required": [
                "expires_in",
                "url"
            ],
            "properties": {
                "expires_in": {
                    "type": "integer",
                    "example": 300
                },
                "url": {
                    "type": "string",
                    "example": "https://bucket.s3.amazonaws.com/path/to/file?signed-params"
                }
            }
        },
...
        "models.ERROR_CODE": {
            "type": "string",
            "enum": [
                "PROFILE_NOT_FOUND",
                "NO_TRANSCRIPT",
                "AUTH_REQUIRED",
                "USAGE_LIMIT_REACHED",
                "USAGE_RESTRICTED"
            ],
            "x-enum-varnames": [
                "PROFILE_NOT_FOUND",
                "NO_TRANSCRIPT",
                "AUTH_REQUIRED",
                "USAGE_LIMIT_REACHED",
                "USAGE_RESTRICTED"
            ]
        },
        "models.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "code": {
                    "enum": [
                        "PROFILE_NOT_FOUND",
                        "NO_TRANSCRIPT",
                        "AUTH_REQUIRED",
                        "USAGE_LIMIT_REACHED",
                        "USAGE_RESTRICTED"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ERROR_CODE"
                        }
                    ],
                    "example": "PROFILE_NOT_FOUND"
                },
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
...
        "models.SpeechToTextRequest": {
            "type": "object",
            "required": [
                "audio_content",
                "language_code"
            ],
            "properties": {
                "audio_content": {
                    "type": "string",
                    "example": "base64-encoded-audio-content"
                },
                "language_code": {
                    "type": "string",
                    "example": "en-US"
                },
                "premium": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SpeechToTextResponse": {
            "type": "object",
            "required": [
                "transcript"
            ],
            "properties": {
                "transcript": {
                    "type": "string",
                    "example": "Hello, how are you?"
                }
            }
        },
...
        "models.TextToSpeechRequest": {
            "type": "object",
            "required": [
                "language_code",
                "text",
                "voice_name"
            ],
            "properties": {
                "language_code": {
                    "type": "string",
                    "example": "en-US"
                },
                "natural": {
                    "type": "boolean",
                    "example": false
                },
                "text": {
                    "type": "string",
                    "example": "Hello, how are you?"
                },
                "voice_name": {
                    "type": "string",
                    "example": "en-US-Standard-A"
                }
            }
        },
        "models.TextToSpeechResponse": {
            "type": "object",
            "required": [
                "audio_content"
            ],
            "properties": {
                "audio_content": {
                    "type": "string",
                    "example": "base64-encoded-audio-content"
                }
            }
        },
...
        "models.TranslateRequest": {
            "type": "object",
            "required": [
                "sentence",
                "source",
                "target"
            ],
            "properties": {
                "sentence": {
                    "type": "string",
                    "example": "Hello, how are you?"
                },
                "source": {
                    "type": "string",
                    "example": "en"
                },
                "target": {
                    "type": "string",
                    "example": "fr"
                }
            }
        },
        "models.TranslateResponse": {
            "type": "object",
            "properties": {
                "sentence": {
                    "type": "string",
                    "example": "Bonjour, comment allez-vous?"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT Authorization header using Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}